[S08P12A809-397] Update: 일기 서비스 리팩토링

+++ back/imfine/src/main/java/com/idle/imfine/service/diary/impl/DiaryServiceImpl.java
@@ -394,8 +394,10 @@ public class DiaryServiceImpl implements DiaryService {
 
         if (diaryPapers.size() != 0) {
             List<Comment> comments = diaryRepository.findDiaryComments(diary);
-            diaryRepository.deleteCommentsHeart(comments);
-            diaryRepository.deleteComments(comments);
+            if (!comments.isEmpty()) {
+                diaryRepository.deleteComments(comments);
+                diaryRepository.deleteCommentsHeart(comments);
+            }
             diaryRepository.deletePaperHasSymptom(diaryPapers);
             diaryRepository.deletePapersHeart(diaryPapers);
             diaryRepository.deletePapers(diary);
